Refactoring: Extract methods. Move class to new file. Remove old code
