+++ packages/language-server/src/completion.ts
@@ -1,14 +1,8 @@
 import {
-  SoqlFromClauseContext,
-  SoqlFromExprsContext,
-  SoqlInnerQueryContext,
   SoqlParser,
   SoqlQueryContext,
-  SoqlSelectClauseContext,
-  SoqlSelectInnerQueryExprContext,
 } from '@salesforce/soql-parser/lib/generated/SoqlParser';
 import { SoqlLexer } from '@salesforce/soql-parser/lib/generated/SoqlLexer';
-import { SoqlParserListener } from '@salesforce/soql-parser/lib/generated/SoqlParserListener';
 import { LowerCasingCharStream } from '@salesforce/soql-parser';
 import {
   CompletionItem,
@@ -16,23 +10,12 @@ import {
   InsertTextFormat,
 } from 'vscode-languageserver';
 
-import {
-  CommonTokenStream,
-  DefaultErrorStrategy,
-  ParserRuleContext,
-  Token,
-  TokenStream,
-} from 'antlr4ts';
-import {
-  ParseTreeWalker,
-  ParseTreeListener,
-  ErrorNode,
-  RuleNode,
-} from 'antlr4ts/tree';
+import { CommonTokenStream, Parser, TokenStream } from 'antlr4ts';
 
 import * as c3 from 'antlr4-c3';
 import { format } from 'util';
 import { SoqlCompletionErrorStrategy } from './completion/SoqlCompletionErrorStrategy';
+import { SoqlQueryExtractor } from './completion/SoqlQueryExtractor';
 
 const SOBJECTS_ITEM_LABEL_PLACEHOLDER = '__SOBJECTS_PLACEHOLDER';
 const SOBJECT_FIELDS_LABEL_PLACEHOLDER = '__SOBJECT_FIELDS_PLACEHOLDER:%s';
@@ -49,6 +32,45 @@ export function completionsFor(
   parser.errorHandler = new SoqlCompletionErrorStrategy();
 
   const parsedQuery = parser.soqlQuery();
+  const core = createC3CompletionEngine(parser);
+  const completionTokenIndex = findCursorTokenIndex(tokenStream, {
+    line,
+    column: column,
+  });
+
+  if (completionTokenIndex === undefined) {
+    console.error(
+      "Couldn't find cursor position on toke stream! Lexer might be skipping some tokens!"
+    );
+    return [];
+  }
+
+  const c3Candidates = core.collectCandidates(
+    completionTokenIndex,
+    parsedQuery
+  );
+  const itemsFromTokens: CompletionItem[] = generateCandidatesFromTokens(
+    c3Candidates.tokens,
+    lexer
+  );
+  const itemsFromRules: CompletionItem[] = generateCandidatesFromRules(
+    c3Candidates.rules,
+    parsedQuery,
+    tokenStream,
+    completionTokenIndex
+  );
+
+  const completionItems = itemsFromTokens.concat(itemsFromRules);
+
+  // If we got no proposals from C3, handle some special cases "manually"
+  if (completionItems.length == 0) {
+    return handleSpecialCases(parsedQuery, tokenStream, completionTokenIndex);
+  }
+
+  return completionItems;
+}
+
+function createC3CompletionEngine(parser: Parser) {
   const core = new c3.CodeCompletionCore(parser);
   core.translateRulesTopDown = true;
   core.ignoredTokens = new Set([
@@ -58,6 +80,7 @@ export function completionsFor(
     SoqlLexer.PLUS,
     SoqlLexer.MINUS,
   ]);
+
   core.preferredRules = new Set([
     SoqlParser.RULE_soqlFromExprs,
     SoqlParser.RULE_soqlFromExpr,
@@ -67,25 +90,97 @@ export function completionsFor(
     SoqlParser.RULE_soqlInteger,
     SoqlParser.RULE_parseReservedForFieldName, // <-- We list it here so that C3 ignores tokens of that rule
   ]);
-  const tokenIndex = findCursorTokenIndex(tokenStream, {
-    line,
-    column: column,
-  });
+  return core;
+}
+const possibleIdentifierPrefix = /[\w]$/;
+export type CursorPosition = { line: number; column: number };
+/**
+  Return the token index for which we want to provide completion candidates,
+  which depends on the cursor possition.
 
-  if (tokenIndex === undefined) {
-    console.error(
-      "Couldn't find cursor position on toke stream! Lexer might be skipping some tokens!"
-    );
-    return [];
+  Examples:
+
+  SELECT id| FROM x     : Cursor touching the previous identifier token:
+                          we want to continue completing that prior token position
+  SELECT id |FROM x     : Cursor NOT touching the previous identifier token:
+                          we want to complete what comes on this new position
+  SELECT id   |  FROM x : Cursor within whitespace block: we want to complete what
+                          comes after the whitespace (we must return a non-WS token index)
+*/
+export function findCursorTokenIndex(
+  tokenStream: TokenStream,
+  cursor: CursorPosition
+) {
+  // NOTE: cursor position is 1-based, while token's charPositionInLine is 0-based
+  const cursorCol = cursor.column - 1;
+  for (let i = 0; i < tokenStream.size; i++) {
+    const t = tokenStream.get(i);
+
+    if (t.line > cursor.line) {
+      return i;
+    }
+
+    let tokenStartCol = t.charPositionInLine;
+    let tokenEndCol = tokenStartCol + (t.text as string).length;
+
+    if (
+      t.line == cursor.line &&
+      (tokenEndCol > cursorCol || t.type === SoqlLexer.EOF)
+    ) {
+      if (
+        i > 0 &&
+        tokenStartCol === cursorCol &&
+        possibleIdentifierPrefix.test(tokenStream.get(i - 1).text as string)
+      ) {
+        return i - 1;
+      } else if (tokenStream.get(i).type === SoqlLexer.WS) {
+        return i + 1;
+      } else return i;
+    }
   }
+}
 
-  const candidatesFromGrammar = core.collectCandidates(tokenIndex, parsedQuery);
-  const completionItems: CompletionItem[] = generateCandidatesFromTokens(
-    candidatesFromGrammar.tokens,
-    lexer
-  );
+function tokenTypeToCandidateString(
+  lexer: SoqlLexer,
+  tokenType: number
+): string {
+  return lexer.vocabulary
+    .getLiteralName(tokenType)
+    ?.toUpperCase()
+    .replace(/^'|'$/g, '') as string;
+}
+function generateCandidatesFromTokens(
+  tokens: Map<number, c3.TokenList>,
+  lexer: SoqlLexer
+): CompletionItem[] {
+  const items: CompletionItem[] = [];
+  for (let [tokenType, followingTokens] of tokens) {
+    const baseKeyword = tokenTypeToCandidateString(lexer, tokenType);
+    if (!baseKeyword) continue;
+    const followingKeywords = followingTokens
+      .map((t) => tokenTypeToCandidateString(lexer, t))
+      .join(' ');
 
-  for (let [ruleId, ruleData] of candidatesFromGrammar.rules) {
+    items.push(
+      newKeywordItem(
+        followingKeywords.length > 0
+          ? baseKeyword + ' ' + followingKeywords
+          : baseKeyword
+      )
+    );
+  }
+  return items;
+}
+
+function generateCandidatesFromRules(
+  c3Rules: Map<number, c3.CandidateRule>,
+  parsedQuery: SoqlQueryContext,
+  tokenStream: TokenStream,
+  tokenIndex: number
+): CompletionItem[] {
+  const completionItems: CompletionItem[] = [];
+
+  for (let [ruleId, ruleData] of c3Rules) {
     switch (ruleId) {
       case SoqlParser.RULE_soqlSelectClause:
         if (tokenIndex <= ruleData.startTokenIndex) {
@@ -147,145 +242,34 @@ export function completionsFor(
         break;
     }
   }
+  return completionItems;
+}
+function handleSpecialCases(
+  parsedQuery: SoqlQueryContext,
+  tokenStream: TokenStream,
+  tokenIndex: number
+): CompletionItem[] {
+  const completionItems: CompletionItem[] = [];
 
-  // If we got no proposals from C3, handle some special cases "manually"
-  if (completionItems.length == 0) {
-    if (
-      isCursorAfter(tokenStream, tokenIndex, [SoqlLexer.SELECT, SoqlLexer.FROM])
-    ) {
-      completionItems.push(newObjectItem(SOBJECTS_ITEM_LABEL_PLACEHOLDER));
-    } else if (
-      isCursorAfter(tokenStream, tokenIndex, [SoqlLexer.COMMA]) &&
-      isCursorBefore(tokenStream, tokenIndex, [SoqlLexer.FROM])
-    ) {
-      const fromSObject =
-        SoqlQueryExtractor.getSObjectFor(parsedQuery, tokenIndex) || 'Object';
-      completionItems.push(
-        newFieldItem(format(SOBJECT_FIELDS_LABEL_PLACEHOLDER, fromSObject))
-      );
-      completionItems.push(
-        newSnippetItem('(SELECT ... FROM ...)', '(SELECT $2 FROM $1)')
-      );
-    }
+  if (
+    isCursorAfter(tokenStream, tokenIndex, [SoqlLexer.SELECT, SoqlLexer.FROM])
+  ) {
+    completionItems.push(newObjectItem(SOBJECTS_ITEM_LABEL_PLACEHOLDER));
+  } else if (
+    isCursorAfter(tokenStream, tokenIndex, [SoqlLexer.COMMA]) &&
+    isCursorBefore(tokenStream, tokenIndex, [SoqlLexer.FROM])
+  ) {
+    const fromSObject =
+      SoqlQueryExtractor.getSObjectFor(parsedQuery, tokenIndex) || 'Object';
+    completionItems.push(
+      newFieldItem(format(SOBJECT_FIELDS_LABEL_PLACEHOLDER, fromSObject))
+    );
+    completionItems.push(
+      newSnippetItem('(SELECT ... FROM ...)', '(SELECT $2 FROM $1)')
+    );
   }
 
   return completionItems;
-}
-
-const possibleIdentifierPrefix = /[\w]$/;
-export type CursorPosition = { line: number; column: number };
-/**
-  Return the token index for which we want to provide completion candidates,
-  which depends on the cursor possition.
-
-  Examples:
-
-  SELECT id| FROM x     : Cursor touching the previous identifier token:
-                          we want to continue completing that prior token position
-  SELECT id |FROM x     : Cursor NOT touching the previous identifier token:
-                          we want to complete what comes on this new position
-  SELECT id   |  FROM x : Cursor within whitespace block: we want to complete what
-                          comes after the whitespace (we must return a non-WS token index)
-*/
-export function findCursorTokenIndex(
-  tokenStream: TokenStream,
-  cursor: CursorPosition
-) {
-  // NOTE: cursor position is 1-based, while token's charPositionInLine is 0-based
-  const cursorCol = cursor.column - 1;
-  for (let i = 0; i < tokenStream.size; i++) {
-    const t = tokenStream.get(i);
-
-    if (t.line > cursor.line) {
-      return i;
-    }
-
-    let tokenStartCol = t.charPositionInLine;
-    let tokenEndCol = tokenStartCol + (t.text as string).length;
-
-    if (
-      t.line == cursor.line &&
-      (tokenEndCol > cursorCol || t.type === SoqlLexer.EOF)
-    ) {
-      if (
-        i > 0 &&
-        tokenStartCol === cursorCol &&
-        possibleIdentifierPrefix.test(tokenStream.get(i - 1).text as string)
-      ) {
-        return i - 1;
-      } else if (tokenStream.get(i).type === SoqlLexer.WS) {
-        return i + 1;
-      } else return i;
-    }
-  }
-}
-
-function newKeywordItem(text: string): CompletionItem {
-  return {
-    label: text,
-    kind: CompletionItemKind.Keyword,
-    insertText: text + ' ',
-  };
-}
-function newFieldItem(text: string): CompletionItem {
-  return {
-    label: text,
-    kind: CompletionItemKind.Field,
-  };
-}
-
-function newNumberItem(text: string): CompletionItem {
-  return {
-    label: text,
-    kind: CompletionItemKind.Constant,
-  };
-}
-function newObjectItem(text: string): CompletionItem {
-  return {
-    label: text,
-    kind: CompletionItemKind.Class,
-  };
-}
-
-function newSnippetItem(label: string, snippet: string): CompletionItem {
-  return {
-    label: label,
-    kind: CompletionItemKind.Snippet,
-    insertText: snippet,
-    insertTextFormat: InsertTextFormat.Snippet,
-  };
-}
-
-function tokenTypeToCandidateString(
-  lexer: SoqlLexer,
-  tokenType: number
-): string {
-  return lexer.vocabulary
-    .getLiteralName(tokenType)
-    ?.toUpperCase()
-    .replace(/^'|'$/g, '') as string;
-}
-function generateCandidatesFromTokens(
-  tokens: Map<number, c3.TokenList>,
-  lexer: SoqlLexer
-): CompletionItem[] {
-  const items: CompletionItem[] = [];
-  for (let [tokenType, followingTokens] of tokens) {
-    const baseKeyword = tokenTypeToCandidateString(lexer, tokenType);
-    if (!baseKeyword) continue;
-    const followingKeywords = followingTokens
-      .map((t) => tokenTypeToCandidateString(lexer, t))
-      .join(' ');
-
-    items.push(
-      newKeywordItem(
-        followingKeywords.length > 0
-          ? baseKeyword + ' ' + followingKeywords
-          : baseKeyword
-      )
-    );
-  }
-  return items;
 }
 
 function isCursorAfter(
@@ -324,209 +308,39 @@ function isCursorBefore(
   }
   return false;
 }
-/***********
 
-interface SelectFromPair {
-  select: Token;
-  from: Token;
-  sobjectName: string;
+function newKeywordItem(text: string): CompletionItem {
+  return {
+    label: text,
+    kind: CompletionItemKind.Keyword,
+    insertText: text + ' ',
+  };
+}
+function newFieldItem(text: string): CompletionItem {
+  return {
+    label: text,
+    kind: CompletionItemKind.Field,
+  };
 }
 
-class FromExpressionExtractor implements SoqlParserListener {
-  sobjectName?: string;
-  closestStartIndex = -1;
-
-  selectsStack: Token[] = [];
-
-  matchedSelectFroms: Array<SelectFromPair> = [];
-
-  constructor(private cursorTokenIndex: number) {}
-
-  static getSObjectFor(
-    parsedQueryTree: SoqlQueryContext,
-    cursorTokenIndex: number
-  ): string | undefined {
-    const listener = new FromExpressionExtractor(cursorTokenIndex);
-    ParseTreeWalker.DEFAULT.walk<SoqlParserListener>(listener, parsedQueryTree);
-    return listener.findSObjectName(cursorTokenIndex);
-  }
-
-  private isInside(pair: SelectFromPair, index: number): boolean {
-    return pair.select.tokenIndex <= index && pair?.from.tokenIndex >= index;
-  }
-  public findSObjectName(index: number): string | undefined {
-    let closest;
-
-    for (let pair of this.matchedSelectFroms) {
-      if (
-        this.isInside(pair, index) &&
-        (!closest ||
-          index - pair.select.tokenIndex < index - closest.select.tokenIndex)
-      ) {
-        closest = pair;
-      }
-    }
-    return closest?.sobjectName;
-  }
-  private foundMatch(fromToken: Token, sobjectName: string) {
-    const selectToken = this.selectsStack.pop();
-    if (selectToken) {
-      this.matchedSelectFroms.push({
-        select: selectToken,
-        from: fromToken,
-        sobjectName: sobjectName,
-      });
-    }
-  }
-
-  visitErrorNode(node: ErrorNode) {
-    if (
-      node.symbol.type === SoqlLexer.FROM &&
-      node.parent &&
-      // node.parent.ruleIndex === SoqlParser.RULE_soqlSelectExpr &&
-      node.parent.childCount >= 2
-    ) {
-      const fromToken = node.symbol;
-      const sobjectName = node.parent.getChild(1).text;
-      this.foundMatch(fromToken, sobjectName);
-    }
-  }
-  enterSoqlInnerQuery(ctx: SoqlInnerQueryContext) {
-    this.selectsStack.push(ctx.start);
-  }
-  exitSoqlFromExprs(ctx: SoqlFromExprsContext) {
-    if (ctx.children && ctx.children.length > 0) {
-      const fromToken = ctx.parent?.start as Token;
-      const sobjectName = ctx.getChild(0).getChild(0).text;
-      this.foundMatch(fromToken, sobjectName);
-    }
-  }
+function newNumberItem(text: string): CompletionItem {
+  return {
+    label: text,
+    kind: CompletionItemKind.Constant,
+  };
+}
+function newObjectItem(text: string): CompletionItem {
+  return {
+    label: text,
+    kind: CompletionItemKind.Class,
+  };
 }
 
-*******/
-
-interface InnerSoqlQuery {
-  soqlInnerQueryNode: SoqlInnerQueryContext;
-  select: Token;
-  from?: Token;
-  sobjectName?: string;
-}
-class SoqlQueryExtractor implements SoqlParserListener {
-  selectsStack: Token[] = [];
-  fromsStack: Array<[Token, string]> = [];
-
-  matchedSelectFroms: Array<InnerSoqlQuery> = [];
-
-  innerSoqlQueries = new Map<number, InnerSoqlQuery>();
-
-  constructor(private cursorTokenIndex: number) {}
-
-  static getSObjectFor(
-    parsedQueryTree: SoqlQueryContext,
-    cursorTokenIndex: number
-  ): string | undefined {
-    const listener = new SoqlQueryExtractor(cursorTokenIndex);
-    ParseTreeWalker.DEFAULT.walk<SoqlParserListener>(listener, parsedQueryTree);
-    return listener.findInnerQuery(cursorTokenIndex)?.sobjectName;
-  }
-  private queryContainsTokenIndex(
-    innerQuery: InnerSoqlQuery,
-    atTokenIndex: number
-  ): boolean {
-    // NOTE: We use the parent node to take into account the enclosing
-    // parentheses (in the case of inner SELECTs), and the whole text until EOF
-    // (for the top-level SELECT). BTW: soqlInnerQueryNode always has a parent.
-    const queryNode = innerQuery.soqlInnerQueryNode.parent
-      ? innerQuery.soqlInnerQueryNode.parent
-      : innerQuery.soqlInnerQueryNode;
-
-    const startIndex = queryNode.start.tokenIndex;
-    const stopIndex = queryNode.stop?.tokenIndex;
-
-    return (
-      atTokenIndex > startIndex && !!stopIndex && atTokenIndex <= stopIndex
-    );
-  }
-
-  private findInnerQuery(atIndex: number): InnerSoqlQuery | undefined {
-    let closestQuery: InnerSoqlQuery | undefined;
-    for (let query of this.innerSoqlQueries.values()) {
-      if (this.queryContainsTokenIndex(query, atIndex)) {
-        closestQuery = query;
-      }
-    }
-    return closestQuery;
-  }
-
-  private findSoqlInnerQueryAncestor(
-    node: RuleNode | undefined
-  ): SoqlInnerQueryContext | undefined {
-    let soqlInnerQueryNode = node;
-    while (
-      soqlInnerQueryNode &&
-      soqlInnerQueryNode.ruleContext.ruleIndex !==
-        SoqlParser.RULE_soqlInnerQuery
-    ) {
-      soqlInnerQueryNode = soqlInnerQueryNode.parent;
-    }
-
-    return soqlInnerQueryNode
-      ? (soqlInnerQueryNode as SoqlInnerQueryContext)
-      : undefined;
-  }
-
-  visitErrorNode(node: ErrorNode) {
-    if (node.parent) {
-      const parentRule = node.parent?.ruleContext.ruleIndex;
-
-      if (
-        parentRule === SoqlParser.RULE_soqlSelectExpr &&
-        node.symbol.type === SoqlLexer.FROM &&
-        node.parent &&
-        node.parent.childCount >= 2
-      ) {
-        const fromToken = node.symbol;
-        const sobjectName = node.parent.getChild(1).text;
-
-        const soqlInnerQueryNode = this.findSoqlInnerQueryAncestor(node.parent);
-        if (soqlInnerQueryNode) {
-          const innerQuery = this.innerSoqlQueries.get(
-            (soqlInnerQueryNode as SoqlInnerQueryContext).start.tokenIndex
-          );
-          if (innerQuery) {
-            innerQuery.from = fromToken;
-            innerQuery.sobjectName = sobjectName;
-          }
-        }
-        // } else if (parentRule === SoqlParser.RULE_soqlSelectClause) {
-        // }
-      }
-    }
-  }
-
-  enterSoqlInnerQuery(ctx: SoqlInnerQueryContext) {
-    this.innerSoqlQueries.set(ctx.start.tokenIndex, {
-      select: ctx.start,
-      soqlInnerQueryNode: ctx,
-    });
-  }
-
-  exitSoqlFromExprs(ctx: SoqlFromExprsContext) {
-    const soqlInnerQueryNode = this.findSoqlInnerQueryAncestor(ctx);
-
-    if (ctx.children && ctx.children.length > 0 && soqlInnerQueryNode) {
-      const fromToken = ctx.parent?.start as Token;
-      const sobjectName = ctx.getChild(0).getChild(0).text;
-      const selectFromPair = this.innerSoqlQueries.get(
-        soqlInnerQueryNode.start.tokenIndex
-      );
-      if (selectFromPair) {
-        selectFromPair.from = fromToken;
-        selectFromPair.sobjectName = sobjectName;
-      }
-    }
-  }
-
-  // exitSoqlInnerQuery(ctx: SoqlInnerQueryContext) {
-  // }
+function newSnippetItem(label: string, snippet: string): CompletionItem {
+  return {
+    label: label,
+    kind: CompletionItemKind.Snippet,
+    insertText: snippet,
+    insertTextFormat: InsertTextFormat.Snippet,
+  };
 }
